generalize resend on Zeroreceipt
