+++ queue/src/Directory.Build.props
@@ -9,7 +9,7 @@
         <PackageReference Include="fiskaltrust.interface" Version="1.3.45">
             <NoWarn>NU1605</NoWarn>
         </PackageReference>
-        <PackageReference Include="fiskaltrust.storage" Version="1.3.45-rc5">
+        <PackageReference Include="fiskaltrust.storage" Version="1.3.45-rc7-23128-62973">
             <NoWarn>NU1605</NoWarn>
         </PackageReference>
     </ItemGroup>
