+++ queue/src/fiskaltrust.Middleware.Contracts/RequestCommands/Factories/IRequestCommandFactory.cs
@@ -1,0 +1,9 @@
+﻿using fiskaltrust.ifPOS.v1;
+
+namespace fiskaltrust.Middleware.Contracts.RequestCommands.Factories
+{
+    public interface IRequestCommandFactory
+    {
+        public RequestCommand Create(ReceiptRequest request);
+    }
+}
