+++ queue/src/fiskaltrust.Middleware.Contracts/RequestCommands/RequestCommand.cs
@@ -11,9 +11,11 @@ namespace fiskaltrust.Middleware.Contracts.RequestCommands
 
         public abstract Task<RequestCommandResponse> ExecuteAsync(ftQueue queue, ReceiptRequest request, ftQueueItem queueItem, bool isResend = false);
 
-        public abstract Task<RequestCommandResponse> ProcessFailedReceiptRequest(ftQueue queue, ftQueueItem queueItem, ReceiptRequest request);
-
         public abstract Task<bool> ReceiptNeedsReprocessing(ftQueue queue, ReceiptRequest request, ftQueueItem queueItem);
+
+        public abstract Task<IQueue> GetIQueue(Guid queueId);
+
+        public abstract Task SaveIQueue(IQueue iQueue);
 
         protected ReceiptResponse CreateReceiptResponse(ftQueue queue, ReceiptRequest request, ftQueueItem queueItem, string ftCashBoxIdentification, long? ftState = null)
         {
@@ -46,6 +48,22 @@ namespace fiskaltrust.Middleware.Contracts.RequestCommands
                 Priority = priority,
                 DataJson = data
             };
+        }
+
+        public async Task<RequestCommandResponse> ProcessFailedReceiptRequest(ftQueue queue, ftQueueItem queueItem, ReceiptRequest request)
+        {
+            var queueIt = await GetIQueue(queue.ftQueueId).ConfigureAwait(false);
+            if (queueIt.SSCDFailCount == 0)
+            {
+                queueIt.SSCDFailMoment = DateTime.UtcNow;
+                queueIt.SSCDFailQueueItemId = queueItem.ftQueueItemId;
+            }
+            queueIt.SSCDFailCount++;
+            await SaveIQueue(queueIt).ConfigureAwait(false);
+            var receiptResponse = CreateReceiptResponse(queue, request, queueItem, queueIt.CashBoxIdentification);
+            receiptResponse.ftState = CountryBaseState | 0x2;
+            receiptResponse.ftStateData = $"Queue is in failed mode. SSCDFailMoment: {queueIt.SSCDFailMoment}, SSCDFailCount: {queueIt.SSCDFailCount}. When connection is established use zeroreceipt for subsequent booking!";
+            return new RequestCommandResponse { ReceiptResponse = receiptResponse };
         }
     }
 }
