+++ queue/src/fiskaltrust.Middleware.Localization.QueueIT/RequestCommands/YearlyClosingReceiptCommand.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
 using fiskaltrust.ifPOS.v1;
-using fiskaltrust.Middleware.Contracts.RequestCommands;
 using fiskaltrust.storage.V0;
 
 namespace fiskaltrust.Middleware.Localization.QueueIT.RequestCommands
@@ -24,6 +23,5 @@ namespace fiskaltrust.Middleware.Localization.QueueIT.RequestCommands
         }
 
         public override Task<bool> ReceiptNeedsReprocessing(ftQueue queue, ReceiptRequest request, ftQueueItem queueItem) => Task.FromResult(false);
-        public override Task<RequestCommandResponse> ProcessFailedReceiptRequest(ftQueue queue, ftQueueItem queueItem, ReceiptRequest request) => throw new NotImplementedException();
     }
 }
