+++ queue/test/Directory.Build.props
@@ -14,7 +14,7 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="fiskaltrust.storage" Version="1.3.45-rc5" />
+    <PackageReference Include="fiskaltrust.storage" Version="1.3.45-rc7-23128-62973" />
     <PackageReference Include="fiskaltrust.interface" Version="1.3.45" />
   </ItemGroup>
 </Project>
