+++ queue/test/fiskaltrust.Middleware.Localization.QueueIT.UnitTest/RequestCommandsTest/PosReceiptCommandTest.cs
@@ -76,7 +76,7 @@ namespace fiskaltrust.Middleware.Localization.QueueIT.UnitTest
             var configRepoMock = new Mock<IConfigurationRepository>();
             configRepoMock.Setup(x => x.GetQueueITAsync(It.IsAny<Guid>())).ReturnsAsync(queueIt);
 
-            var posReceiptCommand = new PosReceiptCommand(desscdMock.Object, new SignatureItemFactoryIT(), Mock.Of<IMiddlewareJournalITRepository>(), configRepoMock.Object, Mock.Of<ILogger<RequestCommand>>());
+            var posReceiptCommand = new PosReceiptCommand(desscdMock.Object, new SignatureItemFactoryIT(), Mock.Of<IMiddlewareJournalITRepository>(), configRepoMock.Object);
 
             var queue = new ftQueue() { ftQueueId = Guid.NewGuid(), ftReceiptNumerator = 5 };
             var queueItem = new ftQueueItem() { ftQueueId = queue.ftQueueId, ftQueueItemId = Guid.NewGuid(), ftQueueRow = 7 };
